fix(api): #5 - applying review
